implemented the remaining part of the reviewServiceImpl

+++ review-service/src/main/java/com/company/MakeMyTrip/review_service/dtos/ReviewSummaryResponse.java
@@ -14,5 +14,5 @@ public class ReviewSummaryResponse {
 
     private Long bookingId;
     private Double averageRating;
-    private Long totalReviews;
+    private Double totalReviews;
 }
